(GH-252) Remove the puppetstrings featureflag
This commit removes the puppet-strings feature flag, and any forking
(if-then-else) that was previously done on the puppet-strings feature-flag.

This commit also moves any puppet-strings feature flags tests to the default
test suite. Thereby making puppet-strings the default code path for everything.
