+++ lib/puppet-languageserver/client_session_state.rb
@@ -48,32 +48,15 @@ module PuppetLanguageServer
 
     # Loaders for object cache information
     def load_default_data!(async = true)
-      if PuppetLanguageServer.featureflag?('puppetstrings')
-        PuppetLanguageServer.log_message(:info, "Loading Default Data via aggregate #{'(Async)' if async}...")
-        if async
-          sidecar_queue.enqueue('default_aggregate', [], false, connection_id)
-          sidecar_queue.enqueue('facts', [], false, connection_id)
-        else
-          sidecar_queue.execute('default_aggregate', [], false, connection_id)
-          sidecar_queue.execute('facts', [], false, connection_id)
-        end
+      PuppetLanguageServer.log_message(:info, "Loading Default Data via aggregate #{'(Async)' if async}...")
+      if async
+        sidecar_queue.enqueue('default_aggregate', [], false, connection_id)
+        sidecar_queue.enqueue('facts', [], false, connection_id)
       else
-        # Order is (somewhat) important here. Try to get the most common information first
-        [
-          { :name => 'Puppet Types', :action => 'default_types' },
-          { :name => 'Facter',       :action => 'facts' },
-          { :name => 'Functions',    :action => 'default_functions' },
-          { :name => 'Classes',      :action => 'default_classes' },
-          { :name => 'DataTypes',    :action => 'default_datatypes' }
-        ].each do |item|
-          PuppetLanguageServer.log_message(:info, "Loading #{item[:name]} #{'(Async)' if async}...")
-          if async
-            sidecar_queue.enqueue(item[:action], [], false, connection_id)
-          else
-            sidecar_queue.execute(item[:action], [], false, connection_id)
-          end
-        end
+        sidecar_queue.execute('default_aggregate', [], false, connection_id)
+        sidecar_queue.execute('facts', [], false, connection_id)
       end
+
       true
     end
 
@@ -87,35 +70,21 @@ module PuppetLanguageServer
         PuppetLanguageServer.log_message(:info, 'Loading static data...')
         load_static_data_impl
       end
+
       true
     end
 
     def load_workspace_data!(async = true)
       return true if documents.store_root_path.nil?
       action_args = ['--local-workspace', documents.store_root_path]
-      if PuppetLanguageServer.featureflag?('puppetstrings')
-        PuppetLanguageServer.log_message(:info, "Loading Workspace Data via aggregate #{'(Async)' if async}...")
-        if async
-          sidecar_queue.enqueue('workspace_aggregate', action_args, false, connection_id)
-        else
-          sidecar_queue.execute('workspace_aggregate', action_args, false, connection_id)
-        end
+      PuppetLanguageServer.log_message(:info, "Loading Workspace Data via aggregate #{'(Async)' if async}...")
+      if async
+        sidecar_queue.enqueue('workspace_aggregate', action_args, false, connection_id)
       else
-        # Order is (somewhat) important here. Try to get the most common information first
-        [
-          { :name => 'Puppet Types', :action => 'workspace_types' },
-          { :name => 'Functions',    :action => 'workspace_functions' },
-          { :name => 'Classes',      :action => 'workspace_classes' },
-          { :name => 'DataTypes',    :action => 'workspace_datatypes' }
-        ].each do |item|
-          PuppetLanguageServer.log_message(:info, "Loading #{item[:name]} #{'(Async)' if async}...")
-          if async
-            sidecar_queue.enqueue(item[:action], action_args, false, connection_id)
-          else
-            sidecar_queue.execute(item[:action], action_args, false, connection_id)
-          end
-        end
+        sidecar_queue.execute('workspace_aggregate', action_args, false, connection_id)
       end
+
+      true
     end
 
     def purge_workspace_data!
