+++ lib/puppet_languageserver_sidecar.rb
@@ -59,11 +59,6 @@ module PuppetLanguageServerSidecar
       flags = options[:flags]
       log_message(:debug, "Detected feature flags [#{options[:flags].join(', ')}]")
 
-      if flags.include?('puppetstrings') && Gem::Version.new(Puppet.version) < Gem::Version.new('5.0.0')
-        # The puppetstrings flag is only valid on Puppet 5.0.0+
-        PuppetEditorServices.log_message(:error, "The feature flag 'puppetstrings' has been specified but it is not capable due to low Puppet version (< 5). Turning off the flag.")
-        flags -= ['puppetstrings']
-      end
       configure_featureflags(flags)
     end
 
@@ -75,17 +70,10 @@ module PuppetLanguageServerSidecar
       sidecar_protocol_extensions
       workspace
       facter_helper
+      puppet_helper
+      puppet_monkey_patches
+      puppet_strings_helper
     ]
-
-    # Load files based on feature flags
-    if featureflag?('puppetstrings')
-      require_list << 'puppet_helper_puppetstrings'
-      require_list << 'puppet_monkey_patches_puppetstrings'
-      require_list << 'puppet_strings_helper'
-    else
-      require_list << 'puppet_helper'
-      require_list << 'puppet_monkey_patches'
-    end
 
     require_list.each do |lib|
       begin
@@ -252,7 +240,10 @@ module PuppetLanguageServerSidecar
   end
 
   def self.execute(options)
-    use_puppet_strings = featureflag?('puppetstrings')
+    unless Gem::Version.new(Puppet.version) >= Gem::Version.new('5.0.0')
+      log_message(:error, 'Only Puppet Version 5.0 and above is supported.')
+      return []
+    end
 
     log_message(:debug, "Executing #{options[:action]} action")
     case options[:action].downcase
@@ -261,43 +252,23 @@ module PuppetLanguageServerSidecar
 
     when 'default_aggregate'
       cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
-      if use_puppet_strings
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache, :object_types => PuppetLanguageServerSidecar::PuppetHelper.available_documentation_types)
-      else
-        create_aggregate(cache)
-      end
+      PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache, :object_types => PuppetLanguageServerSidecar::PuppetHelper.available_documentation_types)
 
     when 'default_classes'
       cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
-      if use_puppet_strings
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache, :object_types => [:class]).classes
-      else
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_classes(cache)
-      end
+      PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache, :object_types => [:class]).classes
 
     when 'default_datatypes'
       cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
-      if use_puppet_strings
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache, :object_types => [:datatype]).datatypes
-      else
-        PuppetLanguageServer::Sidecar::Protocol::PuppetDataTypeList.new
-      end
+      PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache, :object_types => [:datatype]).datatypes
 
     when 'default_functions'
       cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
-      if use_puppet_strings
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache, :object_types => [:function]).functions
-      else
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_functions(cache)
-      end
+      PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache, :object_types => [:function]).functions
 
     when 'default_types'
       cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
-      if use_puppet_strings
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache, :object_types => [:type]).types
-      else
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_types(cache)
-      end
+      PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache, :object_types => [:type]).types
 
     when 'node_graph'
       inject_workspace_as_module || inject_workspace_as_environment
@@ -327,63 +298,37 @@ module PuppetLanguageServerSidecar
     when 'workspace_aggregate'
       return nil unless inject_workspace_as_module || inject_workspace_as_environment
       cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
-      if use_puppet_strings
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache,
-                                                                              :object_types => PuppetLanguageServerSidecar::PuppetHelper.available_documentation_types,
-                                                                              :root_path    => PuppetLanguageServerSidecar::Workspace.root_path)
-      else
-        create_aggregate(PuppetLanguageServerSidecar::Cache::Null.new, PuppetLanguageServerSidecar::Workspace.root_path)
-      end
+      PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache,
+                                                                            :object_types => PuppetLanguageServerSidecar::PuppetHelper.available_documentation_types,
+                                                                            :root_path    => PuppetLanguageServerSidecar::Workspace.root_path)
 
     when 'workspace_classes'
-      null_cache = PuppetLanguageServerSidecar::Cache::Null.new
       return nil unless inject_workspace_as_module || inject_workspace_as_environment
-      if use_puppet_strings
-        cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache,
-                                                                              :object_types => [:class],
-                                                                              :root_path    => PuppetLanguageServerSidecar::Workspace.root_path).classes
-      else
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_classes(null_cache,
-                                                                   :root_path => PuppetLanguageServerSidecar::Workspace.root_path)
-      end
+      cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
+      PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache,
+                                                                            :object_types => [:class],
+                                                                            :root_path    => PuppetLanguageServerSidecar::Workspace.root_path).classes
 
     when 'workspace_datatypes'
       return nil unless inject_workspace_as_module || inject_workspace_as_environment
-      if use_puppet_strings
-        cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache,
-                                                                              :object_types => [:datatype],
-                                                                              :root_path    => PuppetLanguageServerSidecar::Workspace.root_path).datatypes
-      else
-        PuppetLanguageServer::Sidecar::Protocol::PuppetDataTypeList.new
-      end
+      cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
+      PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache,
+                                                                            :object_types => [:datatype],
+                                                                            :root_path    => PuppetLanguageServerSidecar::Workspace.root_path).datatypes
 
     when 'workspace_functions'
-      null_cache = PuppetLanguageServerSidecar::Cache::Null.new
       return nil unless inject_workspace_as_module || inject_workspace_as_environment
-
-      if use_puppet_strings
-        cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache,
-                                                                              :object_types => [:function],
-                                                                              :root_path    => PuppetLanguageServerSidecar::Workspace.root_path).functions
-      else
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_functions(null_cache,
-                                                                     :root_path => PuppetLanguageServerSidecar::Workspace.root_path)
-      end
+      cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
+      PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache,
+                                                                            :object_types => [:function],
+                                                                            :root_path    => PuppetLanguageServerSidecar::Workspace.root_path).functions
 
     when 'workspace_types'
-      null_cache = PuppetLanguageServerSidecar::Cache::Null.new
       return nil unless inject_workspace_as_module || inject_workspace_as_environment
-      if use_puppet_strings
-        cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache,
-                                                                              :object_types => [:type],
-                                                                              :root_path    => PuppetLanguageServerSidecar::Workspace.root_path).types
-      else
-        PuppetLanguageServerSidecar::PuppetHelper.retrieve_types(null_cache)
-      end
+      cache = options[:disable_cache] ? PuppetLanguageServerSidecar::Cache::Null.new : PuppetLanguageServerSidecar::Cache::FileSystem.new
+      PuppetLanguageServerSidecar::PuppetHelper.retrieve_via_puppet_strings(cache,
+                                                                            :object_types => [:type],
+                                                                            :root_path    => PuppetLanguageServerSidecar::Workspace.root_path).types
 
     when 'facts'
       # Can't cache for facts
@@ -397,15 +342,6 @@ module PuppetLanguageServerSidecar
       log_message(:error, "Unknown action #{options[:action]}. Expected one of #{ACTION_LIST}")
     end
   end
-
-  def self.create_aggregate(cache, root_path = nil)
-    result = PuppetLanguageServer::Sidecar::Protocol::AggregateMetadata.new
-    result.concat!(PuppetLanguageServerSidecar::PuppetHelper.retrieve_types(cache, :root_path => root_path))
-    result.concat!(PuppetLanguageServerSidecar::PuppetHelper.retrieve_functions(cache, :root_path => root_path))
-    result.concat!(PuppetLanguageServerSidecar::PuppetHelper.retrieve_classes(cache, :root_path => root_path))
-    result
-  end
-  private_class_method :create_aggregate
 
   def self.output(result, options)
     if options[:output].nil? || options[:output].empty?
