+++ spec/languageserver-sidecar/unit/puppet-languageserver-sidecar/puppet_helper_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe 'PuppetLanguageServerSidecar::PuppetHelper' do
+describe 'PuppetLanguageServerSidecar::PuppetHelper', :if => Gem::Version.new(Puppet.version) >= Gem::Version.new('5.0.0') do
   let(:subject) { PuppetLanguageServerSidecar::PuppetHelper }
   let(:cache) { PuppetLanguageServerSidecar::Cache::Null.new }
 
@@ -13,7 +13,7 @@ describe 'PuppetLanguageServerSidecar::PuppetHelper' do
 
       expect(Puppet).to receive(:settings).and_return(mocked_puppet_settings).at_least(:once)
 
-      result = subject.send(:retrieve_classes, cache, {})
+      result = subject.send(:retrieve_via_puppet_strings, cache, { object_types: [:class] })
     end
   end
 end
